backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
psycopg2-binary
databases[asyncpg]
python-dotenv